backend/requirements.txt
flask
flask-cors
requests
google-generativeai
coqui-tts
numpy
